Move common base database elements under the same directory
