+++ backend/src/database/entities/thumbnail.entity.ts
@@ -1,5 +1,5 @@
 import { Entity, Property } from "@mikro-orm/core";
-import { BaseEntity } from "../../common/entities/base.entity";
+import { BaseEntity } from "../../common/database/base.entity";
 
 @Entity({ tableName: "thumbnails" })
 export class Thumbnail extends BaseEntity {
